+++ R/combineFeatures.R
@@ -1,5 +1,6 @@
 combineFeatures <- function(object,
                             groupBy,
+                            fcol,
                             fun = c("mean",
                                 "median",
                                 "weighted.mean",
@@ -15,6 +16,12 @@ combineFeatures <- function(object,
                             ) {
     if (is.character(fun))
         fun <- match.arg(fun)
+    if (missing(groupBy)) {
+        if (missing(fcol))
+            stop("Require either 'groupBy' or 'fcol'.")
+        stopifnot(fcol %in% fvarLabels(object))
+        groupBy <- fData(object)[, fcol]
+    }
     if (is.list(groupBy)) {
         if (length(groupBy) != nrow(object))
             stop("'length(groupBy)' must be equal to 'nrow(object)': ",