+++ NEWS.md
@@ -7,6 +7,8 @@
 - The MSnSetList class has a new `featureData` slot, accessible with
   `fData` to store metadata for the individial MSnSets of the
   list. `MSnSetList` also not has an `sapply` method. <2018-01-10 Wed>
+- combineFeatures now has a new `fcol` argument (see issue #195)
+  <2018-01-11 Thu>
 
 ## Changes in version 2.5.3
 - MSnSet to/from SummarizedExperiment coercion (contributed by Arne